Extracts donation filter to also add to purchases. Adds specs. #1092
