+++ app/views/purchases/index.html.erb
@@ -33,6 +33,14 @@
             </div>
           <% end %>
         </div><!-- /.row -->
+        <div class="form-group col-lg-3 col-md-4 col-sm-6 col-xs-12">
+          <%= label_tag "from Date" %>
+          <%= date_field :dates, :date_from, class: "form-control", value: @date_from %>
+        </div>
+        <div class="form-group col-lg-3 col-md-4 col-sm-6 col-xs-12">
+          <%= label_tag "to Date" %>
+          <%= date_field :dates, :date_to, class: "form-control", value: @date_to %>
+        </div>
         <div class="row">
           <div class="col-xs-12">
             <%= filter_button %>
